added models spec, lint & prettier config

+++ migrations/sqls/20220804185112-products-table-up.sql
@@ -1,1 +1,1 @@
-CREATE TABLE products (product_id SERIAL PRIMARY KEY, name VARCHAR(100), price NUMERIC(12,2), category VARCHAR(50), user_id INTEGER REFERENCES users(user_id));
+CREATE TABLE products (product_id SERIAL PRIMARY KEY, name VARCHAR(100), price FLOAT, category VARCHAR(50), user_id INTEGER REFERENCES users(user_id));
